tools/gcptranslate/pyproject.toml
[tool.poetry]
name = "translate-api"
version = "0.1.0"
description = ""
authors = ["Yuval Greenfield <ubershmekel@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
google-cloud-translate = "^3.12.0"
google-api-python-client = "^2.97.0"
protobuf = "^4.24.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
